Reorganization of the Separation Logic in MSL. This does not affect users of the Separation Algebras (most of veric for example, or soundness proof in msl/examples/cont/model.v), but it affects the clients of Separation Logic as a "Logic", that is, msl/seplog.v, msl/alg_seplog.v.
The main changes are that there are more axioms for ewand, unfash, and
other things related to subtyping.  Also the Classes in alg_seplog.v have
been reorganized and improved.

Finally, the "cont" case study in msl/examples/cont has been modularized
so that "model.v" uses sep-algs to implement the "logical" interface in
"seplog.v", and then the clients use only the formal logic, not the model.