Updated makefile so that it includes files for new simulation proof.
Solved the universe isse (failure of congruence tactic). The culprit seems
to be type-level let-declarations in the context. Clearing these prior to
invoking congruence solved the problem.

Updated definitions of simulation diagrams so that they accept CoreSemantics' rather than CompcertCoreSems. For Sim_eq, we may use arbitrary type M, for sim_extent and sim_inj we have to instantiate to mem. But the two axioms in CopcertCoreSem, namely corestep_fun and allowed_modifications, were apparently not used for CC_implies_FW or Cminorgenproof. Updated these proofs so that they operate on CoreSem's.
Proof of cminorgen still has a hole wrt initial states.