Handle the parameterization by juicy_ext_spec in the object logic instead of via the module system.
This means:  Make a Class OracleKind that bundles a type with
the juicy_ext_spec over that type.  In semax, replace the former
parameters  {Z}(Hspec: juicy_ext_spec Z)  with just 
{Espec: OracleKind}, an implicit parameter that will often
be supplied by the typeclass system.

This will enable better reasoning about separate compilation and
linking in the logic.