major updates to extension_proof.v to support effect-based formulation of rely-guarantee conditions.  the proof now builds but with a bunch of (hopefully minor) holes (which i'll attempt to fill over the next few days).
one hole, related to pointers received from the environment, may require
an additional condition in the after-external case of the simulation 
diagram.  

the proof has also been updated to eliminate the need for 'stable' 
rely-guarantee simulations, by ensuring that cores that are not 
'focused' -- i.e., actively running in the context of a given extension -- 
are blocked at-external.  this will be true both of the linking semantics
and of our nonpreemptive concurrency semantics.

also: add a coinductive formulation of "compile_safe"

other updates include: 
-removing 'Load loadpath' commands in various places
-probably some other minor stuff...