The "abbreviate" trick.  Now, while doing forward proof in floyd, after each "forward" the Delta, the POSTCONDITION, and the MORE_COMMANDS that come after the first command in a sequence, are pulled out as definitions; and those definitions are to (abbreviate x), where abbreviate is an identity function whose notation is just   "...".
The purpose is twofold:
1.  Put less gunk below the line (and above it) that gets
in the way of seeing what the proof obligation is
2.  Put less gunk below the line that is traversed by the
"rewrite" tactic when it's looking for things to rewrite.

Right now it works pretty well, but it's not perfect
(there's a FIXME in verif_queue.v having to do with
an if statement that is not part of a sequence).