Edit the Makefile so it builds.  The main thing is to comment out these files from the SEPCOMP list:
# extension_proof.v extension_safety.v extension_proof_safety.v
# null_extension.v fs_extension.v linking_extension.v trace_extension.v 


If it is desired to have these files build in the VST release 1.0,
then someone should edit them by (approximately) Wednesday so they
compile, then edit the makefile.  But I'm not sure they need
to be in VST 1.0; it might be fine either way.

All this is now compatible with CompCert 2.0.