In forward.v, clean up the semax_load_tac.  Previously there were three cases:   (Sset _ (Efield (Ederef _ _) _ _))   (Sset _ (Efield _ _ _))   (Sset _ (Ederef _ _))
Now the first case has been eliminated, subsumed into the second case.
This will help when we want to add more support for array subscripting.

What made this slightly difficult is that we need extra work
to eliminate instances of "force_ptr", i.e. that the expression being
dereferenced really is a pointer.  

Next, the store_field_tac really ought to be cleaned up in just the same way.