In res_predicates: Prove superprecise_address_mapsto. This relies on decode_val_uniq, also in the same file. That lemma was not true in CompCert 2.0, but is true (but not proved) in the CompCert trunk (because of changes to not-a-number floats). The proof of decode_val_uniq still has some holes, regarding arithmetic.
In assert_lemmas:
Make the cancel lemma do "eauto with nocore cancel" 
instead of "auto with nocore cancel," so that it can instantiate
unification variables.

In loadstore_lemmas:
Prove semax_load_field_37  and related lemmas.
This avoids forcing the "good" mapsto to the front of the list;
will make symbolic execution work smoother.