Beef up the "normalize" and "entailer" tactics to solve certain goals regarding arithmetic.
1.  computable
Some goals, such as  
x := 8
------
Int.lt (Int.repr 4) (Int.repr x)

are solvable by "compute."  But it's dangerous to call compute,
because it infinite-loops on unsolvable goals.
The new tactic "computable" checks (quickly) for a 
class of goals that compute _can_ solve, then calls compute on them.
This is now part of the "entailer".

2.  repable_signed
Some improvements to this tactic; and now
the "norm" database contains "Int.signed_repr using repable_signed".