update makefile with -slash on coqdep to work with cygwin small update to forward_if tactic
There are some remaining bugs in verif_message and verif_sha on fresh
checkout. There might also still be a bug with external compcert and Cop2.v.

I also don't know if insertion sort should be included in the next build.
It depends on an admitted lemma in list_dt, as well as a few admits in the code.