Added a function, do_computation_equal, for computing on equalities, using the decision procedures built into the reified types.
Also added decision procedures to many of the types that it seemed
like it would be useful to have decision procedures for,
in order to be able to test do_computation_equal.

I still need to prove the correcness of do_computation_equal, but
hopefully most of the heavy lifting has already been done in the
correctness proof of do_computation.