Added alternate names from lseg lemmas, as well as proving first unfolding lemma.
Played around, trying to find bugs in lemma reification and cancellation.
Added two eval hnf statements in hint reification; this was necessary
in order to allow it to successfully process tuples.

I'm noticing there are still issues in the canceler; perhaps eval hnf
statements need to be added there as well? Something weird is going on.
This can be seen in the first goal that I added in reverse_reify_auto.