Filled a few holes.
Oops!  The semax_call rule was unsound, because one of its "admits"
could only be proved with a new premise.  The new premise
checks that,  in   x=f(y,z),
the type of variable x matches the return type of f.

It's easy to check this computationally, and all the proofs
in progs and sha now go through again, without any significant
changes.