CompCert's "decode_val" is not injective, because of a single bit: when 32-bit floats are converted to 64-bit floats, signalling NaNs are converted to quiet NaNs.  Therefore, "mapsto" is not superprecise.  This means that semax_load_37 cannot be derived from semax_load.
The solution, with this check-in, is to remove the proofs
about injectivity of decode_val.  They are all in the
new "superprecise.v".  There are no admits in that file,
but some of the lemmas have an extra premise "wishes_eq_horses",
and are therefore useless.  I preserve them because  it is possible
that future versions of CompCert will have injective decode_val.

In this check-in, the new "semax_load" rule in SeparationLogic.v
is identical to the old "semax_load_37" corollary from floyd/loadstore_lemmas.v.
The old "semax_load" is derivable as a corollary, without needing
"superprecise_mapsto".  But we never had any use for the old
semax_load, we were using semax_load_37 for everything anyway.

The net effect is to remove one "admit".