New notion "precondition_closed", now required for each function.  It states that the precondition of a function-spec is closed w.r.t. any tempvar not mentioned in the formals, and is closed w.r.t. all addressable local vars.
This means there's a new proof obligation for users
when using semax_func_cons, but it's usually solved
completely automatically by the tactic precondition_closed.

All this was necessary to fill a hole in semax_call.
The hole is not filled yet, but now at least it's
more plausible than it was.