1.  Port to CompCert 2.3
2.  Make the definition of match_fdecs more liberal, which permits
  this new rule:

Axiom semax_func_skip: 
  forall {Espec: OracleKind},
   forall V (G: funspecs) fs idf (G': funspecs),
      semax_func V G fs G' ->
      semax_func V G (idf::fs) G'.

That is, if there are extra function-declarations or extern-declarations
in the program that you don't write specifications for, it's OK.

This is extremely useful with CompCert 2.3 that adds many many
externs to the program.  Now there's no need for 
do_builtins, Gtot, or semax_external_FF.  
There's a new tactic, semax_func_skipn, that skips past all
the unused functions; grep in progs/ for examples of its use.