Modify opsems so that the extra work for external calls does not occur as an extra step in Mach, but in the form of rule duplication. Adapted StackingproofEFF.v accordingly.
Corrected the definition of agree and zero_orX in Asmgenproof0EFF.v, so that 
sp and ra blocks can be renamed by the mapping (ie b -> tb, 0 instead of b -> (b,0))

All this was necessary to obtain proofs for the function call cases 
(Mcall, Mtailcall, internalCall,...) in AsmgenproofEFF.v. 

Thus, the proofs of the diagram/effdiagram lemmas in AsmgenproofEFF.v 
are now complete (of course, we still don't support annot etc). 
All remaining holes are in the clauses init/atExternal/afterExternal/Halted.