Split the last component of the tycontext into two.
That is, it was a PTree mapping identifiers to
"global specs" which were either var-types or fun-specs.

Now, it is two separate PTrees.  
The types of vars and functions are in the glob_types
component, and the funspecs of functions are in the glob_specs component.

It's true but UNENFORCED that every function in the glob_specs
is also represented (via type_of_funspec) in the glob_types.