Several improvements to floyd, fixing bugs noticed by Lennart.
1. Make the "cancel" tactic much faster by limiting the rules
in the "Hint Resolve : cancel" database to avoid fully general
unification.  That is, change a bunch of "Hint Resolve"
to "Hint Extern" that limit the match.  This also makes
forward_call' faster, since it uses cancel.

2. check_canonical_funspec now unfolds "MORE_COMMANDS" (or other
   abbreviations) if necessary.

3. slightly more proof automation regarding lvar and eval_var.

4. small performance improvement to Forall_pTree_from_elements,
used in forward_call' tactic.

5. forward_call' now handles calls that are not in a Sequence.

6. init_data2pred' now uses eval_sgvar instead of eval_var,
so that initialized globals that point to other globals work properly.