Some progress on fixing bugs in the new data_at tactics, so that the proofs of verif_queue, verif_nest2 work better.
Problems:
verif_queue: pretty much works, but it uses a LOT more memory,
so that this file can no longer build in 2 gigabytes.

verif_nest2: illustrates a bug in the nested load,
when the nesting level is 2.

Also, changed list_dt so that the listspec now takes
the _name_ of the struct, rather than the struct type,
as an argument.