VeriC Type-check assertion now has type   environ -> mpred instead of   environ->Prop.
This allows the typechecker to inspect the rmap,
to see if we have permission to access certain locations.
It is still the case that eval_expr does NOT access memory,
so can be computed purely on the locals.  However, knowing
about permissions helps us ensure that eval_expr is
correct on pointer comparisons.

The change to actually do pointer-comparisons right
is not in this commit; it'll be in the future.