Many changes to better support permission-shares that are not readable, i.e., that are subshares of Share.Lsh. In particular, the new "permission_block" operator gives permission over a certain address range, without specifying any content; it's a weak form of "memory_block". The new mapsto' operator is like mapsto, except that if the share does not give read access, then it gives a permission_block instead of an actual mapsto. Then, field_at is defined in terms of this mapsto', so that one can now do "field_at sh t path v p" even if "sh" does not give read permission; in which case the value "v" is ignored.   Consequently, field_at no longer implies "readable_share sh".
Part of the motivation for this was to support CompCert 2.5's
new requirements for pointer comparison.  The verif_queue example
is now working again, with a more complicated notion of
a partial-permission list-segment: the permission on the "link"
field does not need to be the same as the permission on the
data fields.