Port to CompCert 2.5 is now done.
The major changes from the last trunk version
are in the treatment of pointer comparison.
Now, for example, (p==NULL)  requires (in the CompCert
operational semantics for eval_expr) access to the memory,
to check whether p is still active; if p has been deallocated,
then this comparison is stuck.

Accommodating this change required major revision to the typechecker;
much of this revision was already done in my previous commit,
but in this commit I still had to make some revisions to the
treatment of unary operators (to accommodate (~p), which now
has a more complex semantics at pointer types).