Remove the "zlist" theory of sublists, and replace it by a less sophisticated (no typeclasses) theory based on firstn/skipn.
Also in this commit:  the value_fits operator, and the guarantee that:
  data_at sh t v p |-- !! value_fits t v.

In particular,  value_fits (tarray t n) v  ->  Zlength v = n.

The next step, after this commit, is to reformulate the
firstn/skipn theory using a "sublist" operator.

Examples in progs/ are all up to date, but sha/ is still
untouched (and broken, i.e. compatible with the master
branch, and CompCert 2.4).