1. forward_call no longer takes an intro_pattern.  Instead, if your postcondition has an EX in it, you use the Intros tactic explicitly after doing forward_call.
2. Intros still moves propositions from the PROP clause above
the line, as before.  But if that proposition is of the form
(x=E), then until now it would automatically do the substitution;
now you must do "subst" manually.

3. start_function no longer does a "simpl" that affects your
function specifications.  Instead, it leaves your funspecs alone,
preserving them unsimplified.

The progs and sha proofs are brought up to date, but not other
proofs (including hmac).