Port VST to CompCert 2.6.
All of progs/* is successfully ported.  Almost no changes were needed
to user proofs, except in relation to external functions.
Still working on sha, hmac, tweetnacl; again, there I expect the
only changes to be related to external functions.

CHANGES:
CompCert 2.6 uses String.string instead of ident (positive) in EF_external
and EF_builtin.  To handle this, veric/ needs a table mapping
strings to their ident.  This table is built automatically
from prog.(prog_defs).  It needs to be passed as an argument to
semax_ext and to add_funspecs.  This is easy to do.

CompCert 2.6 always puts a "return" at the end of a function.
This means that VST's transformation that puts an *extra* return
(if necessary) is no longer needed; hasn't yet been deleted, though.

CompCert 2.6 makes certain functions "Defined" instead of "Qed" in Ctypes.v;
I requested this change September 15, 2015, and Xavier implemented it.

CompCert 2.6's "clightgen" is not broken (as 2.5's was).

I don't recall any other significant changes.  Overall,
this was an easy port to do.  The most difficult part of it was
that CompCert requires ocaml 4.02, which wasn't installed on many
of our computers.