Filled all the holes in veric/initialize.v.
As a consequence, had to change juicy mem alloc_cohere,
to say that the permissions of m_phi are a SUBSET of the permissions of m_dry,
instead of exactly equal.  The difference is only at an
extern global function (Gfun) that is NOT present in the
user's "G" (funspec list).  In such a case, m_phi will
have (NO Share.bot), where the memory will have Nonempty.