1. Undo some of the change in commit 6fcda4a70deafbe124c54560f6ad6cfc90912667. In that commit, I relaxed the access_cohere of juicy_mem, so that m_phi needed only to be a SUBSET of the permissions of m_dry. In this commit, access_cohere is back to matching exactly the (current) permissions of m_dry.
However, in consequence, I needed to find a different way to fix a
problem that the 6fcda4a commit was meant to address:  functions in
the program but not in the funspecs (Gprog) had the wrong permissions.
My fix now is to restrict match_fdecs (internal to semax) and
semax_func (visible to user) so that the funspecs must match the
program's functions exactly.  To make this easier, a new lemma
allows the automatic creation of a vacuous funspec (with precondition FF)
for functions you don't care about.  The function augment_funspecs
can now be used in the construction of Gprog to create all the necessary
vacuous funspecs (see how augment_funspecs is used in the progs/verif_*.v
files).