do not simpl in do_compute_expr,
because that can cause forward to run forever, even if all user-defined
expressions are opaque, because the expression that VST generates by
stepping through C code can become too complex to be efficiently
simpl-able.

Instead of simpl, we now use hnf.
But hnf should not unfold force_val and offset_val, so we have to
fold back these two explicitly.

Moreover, in some cases, the proofs invoking forward (and thus
do_compute_expr) relied on the expression being simplified, but now
these proofs have to simpl themselves, so there are a few updates
to the sha and tweetnacl proofs in this commit.

Tested on floyd progs sha hmac tweetnacl.