add third case to load_tac where user hint is about whole expression
instead of just the root of the expression we load from.

Now load_tac covers the following cases:
1) fully automatic
2) user gives hint how to make root expression nice, or root expression
   already is nice
3) user gives hint how to make full expression nice (unlikely to be
   already nice)

The new case, 3), is less convenient to use, but necessary in cases
where the path written in the syntax is "wrong" and cannot be passed
"as is" to the semax lemma.
An example for this is "p[-1]" in fiddle in progs/load_demo.c.

Moreover, test_legal_nested_efield is removed, because it does not
influence which SEP clause sc_new_instantiate will pick, because when
sc_new_instantiate is called, the address p for which a SEP clause is
to be found as well as the gfs are known, and these two determine which
SEP clause to pick, because they cover disjoint memory areas.

This commit also adds Znth_map rewriting at the end of load_tac,
which "pulls out" the Vint from "map", which allows entailer and other
tactics to be more successful.

TODO/BEWARE: In some corner cases, these changes cause the "lazy" in
solve_load_rule_evaluation to consume unbounded amounts of memory.
This has already happened to me one month ago, and it could be solved
by not calling "subst p" before invoking solve_load_rule_evaluation
(and it is not understandable why this made a difference), but now
I do not find such a "solution".