Doing the comparison in low-level format instead of high-level format
simplifies the equivalence proofs.

The input variables are still in the high-level format, but the
comparison is now done in the low-level format.

Before, we did the following (LL = low-level, HL = high-level):

                convert
  plainHL    ------------>   plainLL
    |                          |
    | encHL                    | encLL
    |                          |
    V           convert        V
  EQUALITY   <------------  cipherLL

Now, we do this:

                convert
  plainHL    ------------>   plainLL
    |                          |
    | encHL                    | encLL
    |                          |
    V           convert        V
  cipherHL   ------------>  EQUALITY

so we do not need any proofs that one conversion is the inverse of the
other, which simplifies the proofs.