in Omega0 (called by myauto/entailer), replace "now" by "solve"
"now t" is defined as "t; easy", and easy does some destruct/inversion
and fails if this does not solve the goal.
It turns out that easy is not needed at all here, but sometimes the
inversion it invokes starts doing beta reduction which takes forever,
so we better do not use easy.