[coq] Update for Coq 8.7.0
VST can't currently compile out-of-the-box with Coq 8.7.0 due to a
missing include. This affects Coq's CI system as they are testing a
branch instead the main VST repository.

The solution is to add the missing include, this however restricts the
out-of-the-box VST to 8.6.1 and 8.7.0+ , removing support for 8.6.0.

IMO this OOTB configuration makes more sense than the current one
requiring a manual compcert update.

We also switch the main Travis testing to 8.7.0 which is slightly
faster and should provide a better testing base.