This commit fixes Issue #117, "Signed integer overflow specification."
Standard C11 treats signed integer overflow as undefined behavior
(not just undefined result, but actually getting stuck),
but CompCert C treats it as 2s-complement modulo arithmetic.
Until now, Verifiable C followed the CompCert C treatment.
However, this meant that C programs proved correct in Verifiable C
could still have wrong behavior when compiled by gcc, and *not*
because of any bug in gcc.  That's bad.

Now, the type-checker for Verifiable C insists that you prove
that signed integer arithmetic does not overflow.  Unsigned
integer arithmetic is not affected by this change.  Only affected
are: add, subtract, multiply, negate.  (Divide was already checked.)

All the proofs in progs, sha, hmac, mailbox, are adjusted.
However: one or two .c programs had to be adjusted to use
unsigned integers instead of signed integers; there are one or two
admits in hmacdrbg because the reseed_counter needs a stronger
invariant (that it's less than Int.max_signed).

Also, in connection with this, there's an extra "simpl" in
the go_lower tactic.  This has very no effect on most proofs,
and no harmful effect anywhere (because go_lower already had a
simpl in it), but in one or two places I had to delete a line
from a proof script because the automation had improved.