entailer tactic no longer does "simpl".
This change was previously made to the entailer! tactic (entbang),
but now it applies to Ltac entailer as well: after performing go_lower,
the tactic does not simplify within the user's PROP and SEP clauses.
This is to avoid blowing up the user's formulas.

This will affect the forward tactic when going through a C return
statement, which applies the entailer tactic: for compatibility,
you may now need to do [simpl] or [simpl; entailer] after the
forward command through a return.

This change may also help resolve issue #224.