Consequence of fixing an ltac inconsistency of intro wrt intros (Coq #11721).
The issue is that names introduced by "intros" behaved as binders for
the continuation of the tactic, but that was not the case of "intro".
As a consequence, in align_mem.v, in a call to "intro t;
type_induction t", the second "t" was considered as a fresh name
rather than as a bound name, later impacting a call to «fresh "t"» in
the body of the Ltac definition for "type_induction".

As a result, a generated name "t0" is now simply "t".

PS: I replaced "intro" with "intros" so as to have a backward
compatible VST PR. Keeping "intro" would be ok, but only one
Coq PR #11721 is merged.