Deprecate floyd.forward.normalize
That is, there is a floyd.seplog_tactics.normalize as well as
a floyd.forward.normalize.  The latter is now considered deprecated.
Whenever the user calls the 'normalize' tactic in a way that
makes use of the forward.normalize features (above and beyond
what seplog_tactics.normalize would do), there will be a warning
message.

Also in this commit:  Remove almost all such uses of 'normalize'
from "make all".