Several minor cleanups; slight incompatibility in forward_if.
Removed several dead tactics in Floyd.
Removed several instances of 'simpl'.
Got rid of do_compute_expr, replacing all its uses by do_compute_expr1
(which is newer and better).  This causes occasional slight incompatibilities
in forward_if:  the symptom is that the new assumption above the line,
expressing that the if-test-expression was true or false, is better
(more processed) than it was before.