Merge %argsassert with %assert.
That is, %argsassert is now just a synonym for %assert.
This simplifies a lot of things, and should make the system
more robust for users, but it has the slight disadvantage
that LOCAL is (inappropriately) active in a PREcondition,
and (vice-versa) PARAMS is active in a loop invariant.
But even when those were more properly distinguished, the
error message was not very helpful if the user got it wrong.

Also, an unrelated tweak in .travis.yml to revert a previous
change that was there only for diagnostic purposes.