Adjust the new "fail 100" error messages for "step" tactic compatibility.
That is:
1. list_solve does not "fail 100", so that users can do "try list_solve".

2. The forward tactics have a "fail softly" mode, enabled by putting
FS := mk_FailSoftly  above the line before calling forward, forward_if,
etc.; and the "step" tactic uses this feature.

3. Some other tactics with "fail 100" have been adjusted.

4. The "entailer" tactic currently does a "fail 100" if the current
proof goal is not an entailment, but it is not clear to me whether
this is a good idea; perhaps we want users to be able to "try entailer!".

With these changes, "make travis" now builds again, but with a change
in sha/verif_hmac_init.v required by point number 4 (that is, it previously
did "try entailer!", and maybe we should continue to permit that).