Several speed improvements in the mkVSU and linkVSUs tactics.
1.  Change the definition of VSU from a sigT to a plain nonconstructive "exists".  It's not at all clear that this helps.

2.  Make several of the tactics used for mkVSU (i.e., mkComponent) and
 linkVSUs faster, by
  a.  more reflective computations;
  b.  judicious use of uconstr instead of constr;
  c.  more careful computation and simplification;
  d.  better algorithms.

It is clear that much of the slowness problems stem from the use of
 -canonical-idents instead of -short-idents  by clightgen.
But in this commit, the VSUpile example continues to be constructed
by -canonical-idents, which gives us a chance to exercise the tactics
and gives us the motivation to improve them.  In a future commit,
we might change back to -short-idents just so the continuous integration
runs faster.