Closes #463.  try_forward_store_union_hack is improved
When two C commands are of the form  "store to one field of a union"
then "load from other field of the union", we already had "if the
fields types are numeric types of the same size, do an automatic
conversion."  But if they were not numeric types of the same size,
then it failed badly.  Now, if they're not numeric types of the same
size, it gives a suggestion to read a (new) chapter of VC.pdf
to learn about how to verify such code.