Adapt to new bitfields from CompCert, as implemented in https://github.com/AbsInt/CompCert/pull/400
Basically, the "member" type describing a field of a struct or union
is now more complicated, so many things in VST must be adapted.
VST does not fully support bitfield-members, only plain members,
but at least it adapts to the new type in the Clight ASTs.

In addition, this adapts to a more recent commit to the
CompCert master branch, moving exportclight/*.v to export/*.v