Change Flocq3 back to Flocq in compcert/*.v
This allows the COMPCERT=bundled setting to work within the
Coq internal continuous integration.

On the other hand, COMPCERT=bundled will not work in the
Coq Platform 2022.04 Coq 8.15.1.  Those (few) users for
whom this is a problem can use the COMPCERT=src_dir build option.

This issue should resolve itself with the Coq 8.15 platform release,
when the entire Flocq3 concept will disappear.