Make `util/make_version` more POSIX-compliant
Fixes #702

I've avoided `&>` as its not POSIX-complaint according to [stack
exchange](https://unix.stackexchange.com/a/590707/45323).

I've also taken the opportunity to simplify slightly the logic that
checks for `git`, relying on the return code of `command -v` rather than
testing that the resulting path exists.

Note that there's no `2>/dev/null` redirect on `command -v`; when the
command does not exist, `command -v` is already silent, and I think we
want noise if `command` itself does not exist or is failing strangely
for some reason.